services/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.10
requests==2.31.0
